+++ WordPress/src/main/java/org/wordpress/android/ui/ActivityLauncher.java
@@ -40,7 +40,7 @@ import org.wordpress.android.ui.accounts.LoginActivity;
 import org.wordpress.android.ui.accounts.LoginEpilogueActivity;
 import org.wordpress.android.ui.accounts.PostSignupInterstitialActivity;
 import org.wordpress.android.ui.accounts.SignupEpilogueActivity;
-import org.wordpress.android.ui.accounts.applicationpassword.test.ApplicationPasswordsListActivity;
+import org.wordpress.android.ui.accounts.applicationpassword.ApplicationPasswordsListActivity;
 import org.wordpress.android.ui.activitylog.detail.ActivityLogDetailActivity;
 import org.wordpress.android.ui.activitylog.list.ActivityLogListActivity;
 import org.wordpress.android.ui.blaze.BlazeFlowSource;
